ResourcesManagerPage API

+++ src/main/java/com/inova/project_manager_api/services/ResourceService.java
@@ -4,6 +4,6 @@ import com.inova.project_manager_api.utils.StandardResponse;
 
 public interface ResourceService {
     StandardResponse searchPotentialResources();
-    StandardResponse getEmployeesNotAllocatedToProject(int projectId);
-    StandardResponse getEmployeesAndProjectsNotAllocatedToProject(int projectId);
+    StandardResponse getEmployeesNotAllocatedToProject(Long projectId);
+    StandardResponse getEmployeesAndProjectsNotAllocatedToProject(Long projectId);
 }
